remove unused fem element: strain smoothing without storing values.
